requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# MongoDB
motor>=3.4.0
pymongo>=4.7.0

# Config
pydantic>=2.7.0
pydantic-settings>=2.2.0
pydantic[email]
python-dotenv>=1.0.0

# LLM & AI
langchain-core>=0.2.0
langchain-anthropic>=0.1.15
langchain-openai>=0.1.8
langchain-google-genai>=1.0.5
langchain-text-splitters>=0.2.0

# NLP / ML
nltk>=3.8.1
scikit-learn>=1.4.2
numpy>=1.26.4
torch>=2.3.0
transformers>=4.41.0

# Document reading
pypdf>=4.2.0
python-docx>=1.1.2

# Web scraping (plágio)
duckduckgo-search>=6.1.7
beautifulsoup4>=4.12.3
requests>=2.32.0

# Segurança
bcrypt>=4.1.3

# Utilitários
python-jose>=3.0.0
passlib>=1.7.4
python-multipart>=0.0.9    
pydantic>=2.7.0
pydantic-settings>=2.2.0
